+++ libs/cli/deepagents_cli/integrations/__init__.py
@@ -1,1 +1,13 @@
 """Sandbox integrations for Deep Agents CLI."""
+
+from deepagents_cli.integrations.sandbox_provider import (
+    SandboxError,
+    SandboxNotFoundError,
+    SandboxProvider,
+)
+
+__all__ = [
+    "SandboxError",
+    "SandboxNotFoundError",
+    "SandboxProvider",
+]
