+++ libs/cli/deepagents_cli/integrations/langsmith.py
@@ -13,8 +13,9 @@ from deepagents.backends.protocol import (
     FileUploadResponse,
     SandboxBackendProtocol,
 )
-from deepagents.backends.sandbox import (
-    BaseSandbox,
+from deepagents.backends.sandbox import BaseSandbox
+
+from deepagents_cli.integrations.sandbox_provider import (
     SandboxListResponse,
     SandboxProvider,
 )
