+++ libs/deepagents/deepagents/backends/sandbox.py
@@ -4,18 +4,16 @@ This module provides a base class that implements all SandboxBackendProtocol
 methods using shell commands executed via execute(). Concrete implementations
 only need to implement the execute() method.
 
-It also defines the SandboxProvider abstract base class for third-party SDK
-implementations to manage sandbox lifecycle (list, create, delete).
+It also defines the BaseSandbox implementation used by the CLI sandboxes.
 """
 
 from __future__ import annotations
 
-import asyncio
 import base64
 import json
 import shlex
 from abc import ABC, abstractmethod
-from typing import Any, Generic, NotRequired, TypeVar
+from typing import NotRequired
 
 from typing_extensions import TypedDict
 
@@ -31,66 +29,7 @@ from deepagents.backends.protocol import (
 )
 
 
-class SandboxError(Exception):
-    """Base exception for sandbox provider operations.
-
-    Provider implementations may raise provider-specific exceptions, but when
-    normalizing errors for callers, prefer raising `SandboxError` (or a subclass)
-    and chain the original exception.
-
-    The recommended pattern is:
-
-    ```python
-    try:
-        ...
-    except Exception as e:
-        raise SandboxError("...") from e
-    ```
-    """
-
-    @property
-    def original_exc(self) -> BaseException | None:
-        """Original exception this error was raised from.
-
-        This is populated when raised using exception chaining:
-
-        ```python
-        raise SandboxError("...") from e
-        ```
-        """
-        return self.__cause__
-
-
-class SandboxNotFoundError(SandboxError):
-    """Raised when a sandbox_id is provided but the sandbox does not exist."""
-
-
-# Type variable for provider-specific metadata
-MetadataT = TypeVar("MetadataT", covariant=True)
-"""Type variable for sandbox metadata.
-
-Providers can define their own TypedDict to specify the structure of sandbox metadata,
-enabling type-safe access to metadata fields.
-
-Example:
-    ```python
-    class ProviderMetadata(TypedDict, total=False):
-        status: Literal["running", "stopped"]
-        created_at: str
-        template: str
-
-    class MyProvider(SandboxProvider[ProviderMetadata]):
-        def list(
-            self, *, cursor=None, **kwargs: Any
-        ) -> SandboxListResponse[ProviderMetadata]:
-            # Extract kwargs as needed
-            status = kwargs.get("status")
-            ...
-    ```
-"""
-
-
-class SandboxInfo(TypedDict, Generic[MetadataT]):
+class SandboxInfo(TypedDict):
     """Metadata for a single sandbox instance.
 
     This lightweight structure is returned from list operations and provides
@@ -128,10 +67,10 @@ class SandboxInfo(TypedDict, Generic[MetadataT]):
     """
 
     sandbox_id: str
-    metadata: NotRequired[MetadataT]
+    metadata: NotRequired[dict[str, object]]
 
 
-class SandboxListResponse(TypedDict, Generic[MetadataT]):
+class SandboxListResponse(TypedDict):
     """Paginated response from a sandbox list operation.
 
     This structure supports cursor-based pagination for efficiently browsing
@@ -158,11 +97,9 @@ class SandboxListResponse(TypedDict, Generic[MetadataT]):
         ```
     """
 
-    items: list[SandboxInfo[MetadataT]]
+    items: list[SandboxInfo]
     cursor: str | None
 
-
-class SandboxProvider(ABC, Generic[MetadataT]):
     """Abstract base class for third-party sandbox provider implementations.
 
     Defines the lifecycle management interface for sandbox providers. Implementations
@@ -211,205 +148,6 @@ class SandboxProvider(ABC, Generic[MetadataT]):
                 self._client.delete(sandbox_id, force=force)
         ```
     """
-
-    @abstractmethod
-    def list(
-        self,
-        *,
-        cursor: str | None = None,
-        **kwargs: Any,
-    ) -> SandboxListResponse[MetadataT]:
-        """List available sandboxes with optional filtering and pagination.
-
-        Args:
-            cursor: Optional continuation token from a previous list() call.
-                Pass None to start from the beginning. The cursor is opaque
-                and provider-specific; clients should not parse or modify it.
-            **kwargs: Provider-specific filter parameters. Implementations should
-                expose these as named keyword-only parameters with defaults for
-                type safety. Common examples include status filters, creation time
-                ranges, template filters, or owner filters.
-
-        Returns:
-            SandboxListResponse containing:
-                - items: List of sandbox metadata for the current page
-                - cursor: Token for next page, or None if this is the last page
-
-        Example:
-            ```python
-            # First page
-            response = provider.list()
-            for sandbox in response["items"]:
-                print(sandbox["sandbox_id"])
-
-            # Next page if available
-            if response["cursor"]:
-                next_response = provider.list(cursor=response["cursor"])
-
-            # With filters (if provider supports them)
-            running = provider.list(status="running")
-            ```
-        """
-
-    @abstractmethod
-    def get_or_create(
-        self,
-        *,
-        sandbox_id: str | None = None,
-        **kwargs: Any,
-    ) -> SandboxBackendProtocol:
-        """Get an existing sandbox or create a new one.
-
-        This method retrieves a connection to an existing sandbox if sandbox_id
-        is provided, or creates a new sandbox instance if sandbox_id is None.
-        The returned object implements SandboxBackendProtocol and can be used
-        for all sandbox operations (execute, read, write, etc.).
-
-        Important: If a sandbox_id is provided but does not exist, this method
-        should raise an error rather than creating a new sandbox. Only when
-        sandbox_id is explicitly None should a new sandbox be created.
-
-        Args:
-            sandbox_id: Unique identifier of an existing sandbox to retrieve.
-                If None, creates a new sandbox instance. The new sandbox's ID
-                can be accessed via the returned object's .id property.
-                If a non-None value is provided but the sandbox doesn't exist,
-                an error will be raised.
-            **kwargs: Provider-specific creation/connection parameters. Implementations
-                should expose these as named keyword-only parameters with defaults
-                for type safety. Common examples include template_id, resource limits,
-                environment variables, or timeout settings.
-
-        Returns:
-            An object implementing SandboxBackendProtocol that can execute
-            commands, read/write files, and perform other sandbox operations.
-
-        Raises:
-            Implementation-specific exceptions for errors such as:
-                - Sandbox not found (if sandbox_id provided but doesn't exist)
-                - Insufficient permissions
-                - Resource limits exceeded
-                - Invalid template or configuration
-
-        Example:
-            ```python
-            # Create a new sandbox
-            sandbox = provider.get_or_create(sandbox_id=None, template_id="python-3.11", timeout_minutes=60)
-            print(sandbox.id)  # "sb_new123"
-
-            # Reconnect to existing sandbox
-            existing = provider.get_or_create(sandbox_id="sb_new123")
-
-            # Use the sandbox
-            result = sandbox.execute("python --version")
-            print(result.output)
-            ```
-        """
-
-    @abstractmethod
-    def delete(
-        self,
-        *,
-        sandbox_id: str,
-        **kwargs: Any,
-    ) -> None:
-        """Delete a sandbox instance.
-
-        This permanently destroys the sandbox and all its associated data.
-        The operation is typically irreversible.
-
-        Idempotency: This method should be idempotent - calling delete on a
-        non-existent sandbox should succeed without raising an error. This makes
-        cleanup code simpler and safe to retry.
-
-        Args:
-            sandbox_id: Unique identifier of the sandbox to delete.
-            **kwargs: Provider-specific deletion options. Implementations should
-                expose these as named keyword-only parameters with defaults for
-                type safety. Common examples include force flags, grace periods,
-                or cleanup options.
-
-        Raises:
-            Implementation-specific exceptions for errors such as:
-                - Insufficient permissions
-                - Sandbox is locked or in use
-                - Network or API errors
-
-        Example:
-            ```python
-            # Simple deletion
-            provider.delete(sandbox_id="sb_123")
-
-            # Safe to call multiple times (idempotent)
-            provider.delete(sandbox_id="sb_123")  # No error even if already deleted
-
-            # With options (if provider supports them)
-            provider.delete(sandbox_id="sb_456", force=True)
-            ```
-        """
-
-    async def alist(
-        self,
-        *,
-        cursor: str | None = None,
-        **kwargs: Any,
-    ) -> SandboxListResponse[MetadataT]:
-        """Async version of list().
-
-        By default, runs the synchronous list() method in a thread pool.
-        Providers can override this for native async implementations.
-
-        Args:
-            cursor: Optional continuation token from a previous list() call.
-            **kwargs: Provider-specific filter parameters.
-
-        Returns:
-            SandboxListResponse containing items and cursor for pagination.
-        """
-        return await asyncio.to_thread(self.list, cursor=cursor, **kwargs)
-
-    async def aget_or_create(
-        self,
-        *,
-        sandbox_id: str | None = None,
-        **kwargs: Any,
-    ) -> SandboxBackendProtocol:
-        """Async version of get_or_create().
-
-        By default, runs the synchronous get_or_create() method in a thread pool.
-        Providers can override this for native async implementations.
-
-        Important: If a sandbox_id is provided but does not exist, this method
-        should raise an error rather than creating a new sandbox. Only when
-        sandbox_id is explicitly None should a new sandbox be created.
-
-        Args:
-            sandbox_id: Unique identifier of an existing sandbox to retrieve.
-                If None, creates a new sandbox instance. If a non-None value
-                is provided but the sandbox doesn't exist, an error will be raised.
-            **kwargs: Provider-specific creation/connection parameters.
-
-        Returns:
-            An object implementing SandboxBackendProtocol.
-        """
-        return await asyncio.to_thread(self.get_or_create, sandbox_id=sandbox_id, **kwargs)
-
-    async def adelete(
-        self,
-        *,
-        sandbox_id: str,
-        **kwargs: Any,
-    ) -> None:
-        """Async version of delete().
-
-        By default, runs the synchronous delete() method in a thread pool.
-        Providers can override this for native async implementations.
-
-        Args:
-            sandbox_id: Unique identifier of the sandbox to delete.
-            **kwargs: Provider-specific deletion options.
-        """
-        await asyncio.to_thread(self.delete, sandbox_id=sandbox_id, **kwargs)
 
 
 _GLOB_COMMAND_TEMPLATE = """python3 -c "
@@ -591,7 +329,6 @@ class BaseSandbox(SandboxBackendProtocol, ABC):
         Returns:
             ExecuteResponse with combined output, exit code, optional signal, and truncation flag.
         """
-        ...
 
     def ls_info(self, path: str) -> list[FileInfo]:
         """Structured listing with file metadata using os.scandir."""
